minor changes in item assignment

+++ application/controllers/Admin.php
@@ -13,6 +13,7 @@ class Admin extends CI_Controller{
         $this->load->helper('paginate');
 
         $this->access['AssetsAccess'] = $this->AssetsAccessList();
+        $this->access['SuppliersAccess'] = $this->SuppliersAccessList();
 
         if(!$this->session->userdata('username')){
             redirect('');
@@ -23,6 +24,12 @@ class Admin extends CI_Controller{
         $user_role = $this->session->userdata('user_role');
         $userAccess = $this->admin_model->request_db_configs($user_role);
         return $userAccess[0];
+    }
+
+    public function SuppliersAccessList() {
+        $user_role = $this->session->userdata('user_role');
+        $userAccess = $this->admin_model->request_db_configs($user_role);
+        return $userAccess[1];
     }
 
     // Load the dashboard.
@@ -238,6 +245,9 @@ class Admin extends CI_Controller{
     }
     // Suppliers - Go to suppliers page.
     public function suppliers($offset = null){
+        if ($this->SuppliersAccessList()->read == 0) {
+            redirect('admin/dashboard');
+        }
        
         $limit = 25;
         if($this->input->get('limit')) {
@@ -280,6 +290,9 @@ class Admin extends CI_Controller{
     }
     // Add Suppliers - Go to add suppliers page.
     public function add_supplier(){
+        if ($this->SuppliersAccessList()->write == 0) {
+            redirect('admin/dashboard');
+        }
         $data['title'] = 'Add Supplier | Admin & Procurement';
         $data['body'] = 'admin/suppliers/add_supplier';
         $data['locations'] = $this->admin_model->list_locations_suppliers();
@@ -291,11 +304,13 @@ class Admin extends CI_Controller{
     }
     // Edit Supplier - Navigates to `Edit Supplier` page
     public function edit_supplier($id = null){  
+        if ($this->SuppliersAccessList()->update == 0) {
+            redirect('admin/dashboard');
+        }
+
         $data['edit'] = $this->admin_model->edit_supplier($id); 
         
         if ($data['edit'] == NULL) {
-            redirect('admin/suppliers');
-        } elseif ($data['edit']->location != $this->session->userdata('location') && $this->session->userdata('user_role') != 'admin') {
             redirect('admin/suppliers');
         }
 
@@ -335,9 +350,14 @@ class Admin extends CI_Controller{
     }
     // Suppliers - Remove supplier
     public function delete_supplier($id){
+        if ($this->SuppliersAccessList()->delete == 0) {
+            redirect('admin/dashboard');
+        }
+
         $data = array(
             'status' => 0
         );
+
         if($this->admin_model->delete_supplier($id,$data)){
             $this->session->set_flashdata('success', '<strong>Success:</strong> Supplier removal was successful.');
             redirect('admin/suppliers');
@@ -416,13 +436,14 @@ class Admin extends CI_Controller{
         $data['title'] = 'Access Control List | Admin & Procurement';
         $data['body'] = 'admin/ACL/acl';
         $data['assets_access_list'] = $this->admin_model->component_access_list('assets');
+        $data['suppliers_access_list'] = $this->admin_model->component_access_list('suppliers');
         $data['acl_page'] = true;
         $data['breadcrumb'] = array("Access Control List");
         
         $this->load->view('admin/commons/new_template', $data);
     } 
     // Form Logic for Assets Access on ACL Page
-    public function update_asset_access() {
+    public function update_access_list() {
         $asset_read = $this->input->post('read'); // read[1][1], read[1][2], read[1][3]
         $asset_write = $this->input->post('write'); 
         $asset_update = $this->input->post('update'); 
@@ -832,13 +853,20 @@ class Admin extends CI_Controller{
         if ($this->AssetsAccessList()->read == 0) {
             redirect('admin/dashboard');
         }
-        $limit = 10;
+
+        $limit = 25;
+        if($this->input->get('limit')) {
+            $limit = $this->input->get('limit');
+        }
+
         if(!empty($offset)){
-            $this->uri->segment(3);
-        } 
+            $config['uri_segment'] = 3;
+        }
+    
         $this->load->library('pagination');
         $url = base_url('admin/asset_register');
         $rowscount = $this->admin_model->count_assets();
+
         $config['base_url'] = $url;
         $config['total_rows'] = $rowscount;
         $config['per_page'] = $limit;
@@ -851,8 +879,9 @@ class Admin extends CI_Controller{
         $config['next_link'] = 'Next';
         $config['last_link'] = 'Last';
         $config['attributes'] = array('class' => 'pagination-link');
-        $this->pagination->initialize($config);  
-        
+        $config['reuse_query_string'] = true;
+        $this->pagination->initialize($config);
+               
         $data['title'] = 'Asset Register | Admin & Procurement';
         $data['body'] = 'admin/asset-register';
         $data['assets'] = $this->admin_model->get_assets($limit, $offset);
@@ -870,6 +899,7 @@ class Admin extends CI_Controller{
         $data['breadcrumb'] = array("admin/asset_register" => "Assets List", "Add Asset");
         $data['add_asset'] = true;
         $data['locations'] = $this->admin_model->get_item_location();
+        $data['categories'] = $this->admin_model->get_item_categories();
         $this->load->view('admin/commons/new_template', $data);
     }
     // Asset detail
@@ -899,7 +929,7 @@ class Admin extends CI_Controller{
                 'user' => $this->input->post('user'),
                 'remarks' => $this->input->post('remarks'), 
                 'created_at' => date('Y-m-d')
-            );
+            ); 
             if($this->admin_model->add_item($data)){
                 $this->session->set_flashdata('success', '<strong>Success! </strong>Item was added successfully.');
                 redirect('admin/asset_register');
